#  Some micro-optimizations of the trajectory iterators